// alfcomponents/enums/alf-position.enum.ts
/**
 * Enum para posicionamiento y alineación genérica (Backgrounds y Posiciones relativas)
 * Se enfoca en las 9 posiciones estándar y desplazamientos básicos.
 */
export enum AlfPositionEnum {
  Auto = 'auto',
  Center = 'center',
  Top = 'top',
  Bottom = 'bottom',
  Left = 'left',
  Right = 'right',
  TopLeft = 'top left',
  TopCenter = 'top center',
  TopRight = 'top right',
  CenterLeft = 'center left',
  CenterCenter = 'center center',
  CenterRight = 'center right',
  BottomLeft = 'bottom left',
  BottomCenter = 'bottom center',
  BottomRight = 'bottom right',
}